% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikemap.R
\name{spikelabel}
\alias{spikelabel}
\title{Get Spike Tips}
\usage{
spikelabel(x, var, inches = 3, fixmax)
}
\arguments{
\item{x}{an sf object, a simple feature collection.}

\item{var}{name of the numeric field in x to plot.}

\item{inches}{size of the tallest spike, in inches.}

\item{fixmax}{value of the tallest spike (see Details).}
}
\value{
An sf POINT layer of spikes tips is returned
}
\description{
Get spikes' tips, may be usefull for labels
}
\examples{
library(sf)
com <- st_read(system.file("gpkg/com.gpkg", package="spikemap"))
plot(st_geometry(com))
spikemap(x = com, var = "pop")
tips <- spikelabel(x = com, var = "pop")
plot(st_geometry(tips), add = TRUE)
}
