% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikemap.R
\name{spikemap}
\alias{spikemap}
\title{Plot a Spike Map}
\usage{
spikemap(
  x,
  var,
  inches = 3,
  width = 0.02,
  fixmax,
  col = "white",
  border = "red",
  lwd = 0.5,
  legend.pos = "bottomleft",
  legend.title.txt = var,
  legend.title.cex = 0.8,
  legend.values.cex = 0.6,
  legend.values.rnd = 0,
  add = TRUE
)
}
\arguments{
\item{x}{an sf object, a simple feature collection.}

\item{var}{name of the numeric field in x to plot.}

\item{inches}{size of the tallest spike, in inches.}

\item{width}{width of the spike's base.}

\item{fixmax}{value of the tallest spike (see Details).}

\item{col}{filling color of the spikes.}

\item{border}{border color of the spikes.}

\item{lwd}{width of spikes borders.}

\item{legend.pos}{position of the legend, "bottomright", or a
vector of two coordinates in map units (c(x, y)). If
legend.pos is "n" then the legend is not plotted.}

\item{legend.title.txt}{title of the legend.}

\item{legend.title.cex}{size of the legend title.}

\item{legend.values.cex}{size of the values in the legend.}

\item{legend.values.rnd}{number of decimal places of the values
displayed in the legend.}

\item{add}{whether to add the layer to an existing plot (TRUE) or
not (FALSE).}
}
\description{
Plot a layer of spikes on a map.
}
\details{
Two maps with the same inches and fixmax parameters will be comparable.
}
\examples{
library(sf)
com <- st_read(system.file("gpkg/com.gpkg", package="spikemap"))
plot(st_geometry(com))
spikemap(x = com, var = "pop")
}
